requirements.txt
anyio
uvicorn
fastapi
sounddevice
git+https://github.com/wumtdev/wauxio.git
git+https://github.com/wumtdev/wsignals.git
aiosqlite
SQLAlchemy
APScheduler
pyjwt
passlib[bcrypt]
Jinja2
